# frontend/openapi/Sabro.API.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Sabro.API | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/play/meltho/today": {
      "get": {
        "tags": [
          "Play"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MelthoPuzzleDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MelthoPuzzleDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MelthoPuzzleDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/play/mno/today": {
      "get": {
        "tags": [
          "Play"
        ],
        "parameters": [
          {
            "name": "difficulty",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/MnoDifficulty"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MnoPuzzleDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MnoPuzzleDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MnoPuzzleDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/play/meltho/library": {
      "get": {
        "tags": [
          "Play"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/LibrarySort"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortDirection"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfMelthoLibraryEntryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfMelthoLibraryEntryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfMelthoLibraryEntryDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/play/meltho/library/{lexiconEntryId}": {
      "get": {
        "tags": [
          "Play"
        ],
        "parameters": [
          {
            "name": "lexiconEntryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MelthoLibraryDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MelthoLibraryDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MelthoLibraryDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/play/results": {
      "post": {
        "tags": [
          "Play"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordGameResultRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordGameResultRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RecordGameResultRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GameResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResultDto"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GameResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/play/results/me": {
      "get": {
        "tags": [
          "Play"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "gameId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfGameResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfGameResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfGameResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GameResultDto": {
        "required": [
          "id",
          "logtoUserId",
          "gameId",
          "playedOn",
          "solved",
          "attempts",
          "detailJson",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "logtoUserId": {
            "type": "string"
          },
          "gameId": {
            "type": "string"
          },
          "playedOn": {
            "type": "string",
            "format": "date"
          },
          "solved": {
            "type": "boolean"
          },
          "attempts": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detailJson": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "LetterHardening": {
        "enum": [
          "Qushoyo",
          "Rukkokho"
        ],
        "type": "string"
      },
...
      "LibrarySort": {
        "enum": [
          "Recent",
          "Alphabetical",
          "Length"
        ],
        "type": "string",
        "default": 0
      },
...
      "MelthoLibraryDetailDto": {
        "required": [
          "lexiconEntryId",
          "syriacUnvocalized",
          "syriacVocalized",
          "sblTransliteration",
          "transliterationVariants",
          "grammaticalCategory",
          "morphology",
          "playableLength",
          "root",
          "meanings",
          "composition",
          "playedOn",
          "pronunciationAudioUrl"
        ],
        "type": "object",
        "properties": {
          "lexiconEntryId": {
            "type": "string",
            "format": "uuid"
          },
          "syriacUnvocalized": {
            "type": "string"
          },
          "syriacVocalized": {
            "type": [
              "null",
              "string"
            ]
          },
          "sblTransliteration": {
            "type": [
              "null",
              "string"
            ]
          },
          "transliterationVariants": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "grammaticalCategory": {
            "type": "string"
          },
          "morphology": {
            "type": [
              "null",
              "string"
            ]
          },
          "playableLength": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "root": {
            "type": [
              "null",
              "string"
            ]
          },
          "meanings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MelthoPuzzleMeaningDto"
            }
          },
          "composition": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyriacLetter"
            }
          },
          "playedOn": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date"
            }
          },
          "pronunciationAudioUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "MelthoLibraryEntryDto": {
        "required": [
          "lastPlayedOn",
          "lexiconEntryId",
          "syriacUnvocalized",
          "syriacVocalized",
          "sblTransliteration",
          "playableLength",
          "timesPlayed",
          "meanings"
        ],
        "type": "object",
        "properties": {
          "lastPlayedOn": {
            "type": "string",
            "format": "date"
          },
          "lexiconEntryId": {
            "type": "string",
            "format": "uuid"
          },
          "syriacUnvocalized": {
            "type": "string"
          },
          "syriacVocalized": {
            "type": [
              "null",
              "string"
            ]
          },
          "sblTransliteration": {
            "type": [
              "null",
              "string"
            ]
          },
          "playableLength": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "timesPlayed": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "meanings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MelthoPuzzleMeaningDto"
            }
          }
        }
      },
      "MelthoPuzzleDto": {
        "required": [
          "date",
          "lexiconEntryId",
          "syriacUnvocalized",
          "syriacVocalized",
          "sblTransliteration",
          "grammaticalCategory",
          "playableLength",
          "meanings"
        ],
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "lexiconEntryId": {
            "type": "string",
            "format": "uuid"
          },
          "syriacUnvocalized": {
            "type": "string"
          },
          "syriacVocalized": {
            "type": [
              "null",
              "string"
            ]
          },
          "sblTransliteration": {
            "type": [
              "null",
              "string"
            ]
          },
          "grammaticalCategory": {
            "type": "string"
          },
          "playableLength": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "meanings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MelthoPuzzleMeaningDto"
            }
          }
        }
      },
      "MelthoPuzzleMeaningDto": {
        "required": [
          "language",
          "text"
        ],
        "type": "object",
        "properties": {
          "language": {
            "type": "string"
          },
          "text": {
            "type": "string"
          }
        }
      },
      "MnoDifficulty": {
        "enum": [
          "Beginner",
          "Easy",
          "Normal",
          "Hard",
          "Extreme"
        ],
        "type": "string",
        "default": 2
      },
      "MnoPuzzleDto": {
        "required": [
          "date",
          "difficulty",
          "target",
          "tileCount",
          "expression",
          "tileForm"
        ],
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "difficulty": {
            "type": "string"
          },
          "target": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "tileCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "expression": {
            "type": "string"
          },
          "tileForm": {
            "type": "string"
          }
        }
      },
...
      "PagedResultOfGameResultDto": {
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameResultDto"
            }
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "PagedResultOfMelthoLibraryEntryDto": {
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MelthoLibraryEntryDto"
            }
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RecordGameResultRequest": {
        "required": [
          "gameId",
          "playedOn",
          "solved",
          "attempts",
          "detailJson"
        ],
        "type": "object",
        "properties": {
          "gameId": {
            "type": "string"
          },
          "playedOn": {
            "type": "string",
            "format": "date"
          },
          "solved": {
            "type": "boolean"
          },
          "attempts": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detailJson": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SortDirection": {
        "enum": [
          "Ascending",
          "Descending"
        ],
        "type": "string"
      },
...
      "SyriacLetter": {
        "required": [
          "letter",
          "code",
          "vowel",
          "isBegadkephat",
          "hardening",
          "hardeningSource"
        ],
        "type": "object",
        "properties": {
          "letter": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/SyriacLetterCode"
          },
          "vowel": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SyriacVowel"
              }
            ]
          },
          "isBegadkephat": {
            "type": "boolean"
          },
          "hardening": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LetterHardening"
              }
            ]
          },
          "hardeningSource": {
            "$ref": "#/components/schemas/HardeningSource"
          }
        }
      },
...
      "SyriacVowel": {
        "enum": [
          "Pthaha",
          "Zqapha",
          "Rbasa",
          "Zlama",
          "Hbasa",
          "Esasa",
          "Rwaha"
        ],
        "type": "string"
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
...
}